test: extract step-config tests; add DefaultOption and Timeout last-write scenarios
